+++ app/src/main/java/ca/appolizer/outreach/MainActivity.java
@@ -22,8 +22,8 @@ import ca.appolizer.outreach.databinding.ActivityMainBinding;
 import ca.appolizer.outreach.model.Student;
 import ca.appolizer.outreach.model.User;
 import ca.appolizer.outreach.ui.job.JobFragment;
+
 import ca.appolizer.outreach.ui.profile.ProfileActivity;
-import ca.appolizer.outreach.ui.profile.ProfileMainActivity;
 
 public class MainActivity extends AppCompatActivity {
     public static final String SHARED_PREFERENCES = "profile_info";
@@ -134,15 +134,11 @@ public class MainActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         switch (item.getItemId()) {
             case R.id.action_profile:
-                Intent intent = new Intent(this, ProfileActivity.class);
-                intent.putExtra("id", id);
-                intent.putExtra("token", token);
-                intent.putExtra("student_profile", user);
-                startActivity(intent);
+                Intent settingIntent = new Intent(this, ProfileActivity.class);
+                startActivity(settingIntent);
                 break;
             case R.id.action_section:
-                Intent settingIntent = new Intent(this, ProfileMainActivity.class);
-                startActivity(settingIntent);
+
                 break;
         }
         return super.onOptionsItemSelected(item);
